watch-query-function/requirements.txt
google-cloud-pubsub==2.18.4
psycopg2-binary==2.9.10
google-auth==2.23.4
google-api-python-client==2.108.0
functions-framework==3.5.0